refactor(internal): move board utils into dedicated module
