+++ a3p-integration/proposals/z:acceptance/test/test-lib/rpc.js
@@ -14,7 +14,7 @@
 import {
   boardSlottingMarshaller,
   makeBoardRemote,
-} from '@agoric/internal/src/marshal.js';
+} from '@agoric/internal/src/marshal/board-client-utils.js';
 import { E, Far } from '@endo/far';
 import { Fail } from '@endo/errors';
