+++ packages/vats/tools/board-utils.js
@@ -5,10 +5,10 @@
 
 /**
  * @typedef {{
- *   brand: import('@agoric/internal/src/marshal.js').BoardRemote;
+ *   brand: import('@agoric/internal/src/marshal/board-client-utils.js').BoardRemote;
  *   denom: string;
  *   displayInfo: DisplayInfo;
- *   issuer: import('@agoric/internal/src/marshal.js').BoardRemote;
+ *   issuer: import('@agoric/internal/src/marshal/board-client-utils.js').BoardRemote;
  *   issuerName: string;
  *   proposedName: string;
  * }} VBankAssetDetail
@@ -17,7 +17,7 @@
  * @typedef {{
  *   brand: Record<
  *     string,
- *     import('@agoric/internal/src/marshal.js').BoardRemote
+ *     import('@agoric/internal/src/marshal/board-client-utils.js').BoardRemote
  *   >;
  *   instance: Record<string, Instance>;
  *   installation: Record<string, Installation>;
@@ -30,12 +30,12 @@
 import {
   slotToBoardRemote,
   unmarshalFromVstorage,
-} from '@agoric/internal/src/marshal.js';
+} from '@agoric/internal/src/marshal/board-client-utils.js';
 import { makeScalarBigMapStore } from '@agoric/vat-data';
 import { makeMarshal } from '@endo/marshal';
 import { prepareBoardKit } from '../src/lib-board.js';
 
-export * from '@agoric/internal/src/marshal.js';
+export * from '@agoric/internal/src/marshal/board-client-utils.js';
 
 /**
  * @param {import('@agoric/internal/src/storage-test-utils.js').FakeStorageKit} fakeStorageKit
